Sync script works when there not separate provR and rdt files.

+++ sync-repos.sh
@@ -61,7 +61,7 @@ function is_current {
   return 1
 }
 
-# Copy just the files that we want in the provR repository.  Some
+# Copyz just the files that we want in the provR repository.  Some
 # files need to be renamed when they are copied to remove _prov
 # from the name. 
 function copy_provR_files {
@@ -69,15 +69,33 @@ function copy_provR_files {
   
   # Copy the files we want to provR
   rsync -tv --perms inst/CITATION ../provR_test/inst/
-  rsync -tv --perms DESCRIPTION_prov ../provR_test/
-  mv ../rdt_test/DESCRIPTION_prov ../rdt_test/DESCRIPTION
-  rsync -tv --perms NAMESPACE_prov ../provR_test/
-  mv ../rdt_test/NAMESPACE_prov ../rdt_test/NAMESPACE
+  if [ -e DESCRIPTION_prov ] 
+  then
+    rsync -tv --perms DESCRIPTION_prov ../provR_test/DESCRIPTION
+  else 
+    rsync -tv --perms DESCRIPTION ../provR_test/DESCRIPTION
+  fi
+  if [ -e NAMESPACE_prov ] 
+  then
+    rsync -tv --perms NAMESPACE_prov ../provR_test/NAMESPACE
+  else
+    rsync -tv --perms NAMESPACE ../provR_test/NAMESPACE
+  fi
   rsync -tv --perms LICENSE ../provR_test/
-  rsync -tv --perms README_prov.md ../provR_test/
+  if [ -e README_prov.MD ] 
+  then
+    rsync -tv --perms README_prov.md ../provR_test/README.md
+  else
+    rsync -tv --perms README.md ../provR_test/README.md
+  fi
   rsync -tv --perms .Rbuildignore ../provR_test/
   rsync -rtv --del --exclude "*_rdt.R" --exclude "DDGCheckpoint.R" --perms R ../provR_test/
-  rsync -rtv --del --perms man_prov/* ../provR_test/man
+  if [ -e man_prov ] 
+  then
+    rsync -rtv --del --perms man_prov/* ../provR_test/man
+  else 
+    rsync -rtv --del --perms man/* ../provR_test/man
+  fi
   rsync -rtv --del --exclude "*_rdt*" --perms tests ../provR_test/tests
 }
 
@@ -89,15 +107,33 @@ function copy_rdt_files {
 
   # Copy the files we want to rdt
   rsync -tv --perms inst/CITATION ../rdt_test/inst/
-  rsync -tv --perms DESCRIPTION_rdt ../rdt_test/
-  mv ../rdt_test/DESCRIPTION_RDT ../rdt_test/DESCRIPTION
-  rsync -tv --perms NAMESPACE_rdt ../rdt_test/
-  mv ../rdt_test/NAMESPACE_RDT ../rdt_test/NAMESPACE
+  if [ -e DESCRIPTION_rdt ] 
+  then
+    rsync -tv --perms DESCRIPTION_rdt ../rdt_test/DESCRIPTION
+  else 
+    rsync -tv --perms DESCRIPTION ../rdt_test/DESCRIPTION
+  fi
+  if [ -e NAMESPACE_rdt ] 
+  then
+    rsync -tv --perms NAMESPACE_rdt ../rdt_test/NAMESPACE
+  else 
+    rsync -tv --perms NAMESPACE ../rdt_test/NAMESPACE
+  fi
   rsync -tv --perms LICENSE ../rdt_test/
-  rsync -tv --perms README_rdt.md ../rdt_test/
+  if [ -e README_rdt.MD ] 
+  then
+    rsync -tv --perms README_rdt.md ../rdt_test/README.md
+  else
+    rsync -tv --perms README.md ../rdt_test/README.md
+  fi
   rsync -tv --perms .Rbuildignore ../rdt_test/
   rsync -rtv --del --exclude "*_prov.R" --exclude "DDGCheckpoint.R" --perms R ../rdt_test/
-  rsync -rtv --del --perms man_prov/* ../rdt_test/man
+  if [ -e man_rdt ] 
+  then
+    rsync -rtv --del --perms man_rdt/* ../rdt_test/man
+  else
+    rsync -rtv --del --perms man/* ../rdt_test/man
+  fi
   rsync -rtv --del --exclude "*_prov*" --perms tests ../rdt_test/tests
 }
